Added compute type option

+++ config.py
@@ -23,7 +23,7 @@ class Config:
         # cannot be changed because you would need
         # to install the appropriate dependencies for each platform
         self.device = "cpu"
-        self.compute_type = "int8"
+        self.compute_type = "float16"
         self.batch_size = 8
         # this isn't used in the code but is here for reference/documentation purposes
         self.supported_languages_shorthand = [
